HBX-1665: Move class 'org.hibernate.tool.hbmlint.IssueCollector' to 'org.hibernate.tool.internal.export.lint.IssueCollector'
