--- orm/src/main/java/org/hibernate/tool/hbmlint/IssueCollector.java
@@ -1,7 +1,0 @@
-package org.hibernate.tool.hbmlint;
-
-public interface IssueCollector {
-
-	public abstract void reportIssue(Issue analyze);
-
-}
